Add support for system C11 threads, e.g. from alpine/musl (#1998)
