+++ src/tinycthread.c
@@ -23,9 +23,9 @@ freely, subject to the following restrictions:
 */
 
 #include "rd.h"
-#include "rdtime.h"
-#include "tinycthread.h"
 #include <stdlib.h>
+
+#if !WITH_C11THREADS
 
 /* Platform specific includes */
 #if defined(_TTHREAD_POSIX_)
@@ -387,7 +387,7 @@ int cnd_broadcast(cnd_t *cond)
 }
 
 #if defined(_TTHREAD_WIN32_)
-static int _cnd_timedwait_win32(cnd_t *cond, mtx_t *mtx, DWORD timeout)
+int _cnd_timedwait_win32(cnd_t *cond, mtx_t *mtx, DWORD timeout)
 {
   int result, lastWaiter;
 
@@ -476,56 +476,6 @@ int cnd_timedwait(cnd_t *cond, mtx_t *mtx, const struct timespec *ts)
 }
 
 
-int cnd_timedwait_ms(cnd_t *cnd, mtx_t *mtx, int timeout_ms) {
-  if (timeout_ms == -1 /* INFINITE*/)
-    return cnd_wait(cnd, mtx);
-#if defined(_TTHREAD_WIN32_)
-	return _cnd_timedwait_win32(cnd, mtx, (DWORD)timeout_ms);
-#else
-  int ret;
-	struct timeval tv;
-	struct timespec ts;
-
-	gettimeofday(&tv, NULL);
-  ts.tv_sec = tv.tv_sec;
-  ts.tv_nsec = tv.tv_usec * 1000;
-
-	ts.tv_sec  += timeout_ms / 1000;
-	ts.tv_nsec += (timeout_ms % 1000) * 1000000;
-
-	if (ts.tv_nsec >= 1000000000) {
-		ts.tv_sec++;
-		ts.tv_nsec -= 1000000000;
-	}
-
-  ret = pthread_cond_timedwait(cnd, mtx, &ts);
-  if (ret == ETIMEDOUT)
-  {
-    return thrd_timedout;
-  }
-  return ret == 0 ? thrd_success : thrd_error;
-#endif
-}
-
-int cnd_timedwait_msp (cnd_t *cnd, mtx_t *mtx, int *timeout_msp) {
-        rd_ts_t pre = rd_clock();
-        int r;
-        r = cnd_timedwait_ms(cnd, mtx, *timeout_msp);
-        if (r != thrd_timedout) {
-                /* Subtract spent time */
-                (*timeout_msp) -= (int)(rd_clock()-pre) / 1000;
-        }
-        return r;
-}
-
-int cnd_timedwait_abs (cnd_t *cnd, mtx_t *mtx, const struct timespec *tspec) {
-        if (tspec->tv_sec == RD_POLL_INFINITE)
-                return cnd_wait(cnd, mtx);
-        else if (tspec->tv_sec == RD_POLL_NOWAIT)
-                return thrd_timedout;
-
-        return cnd_timedwait(cnd, mtx, tspec);
-}
 
 #if defined(_TTHREAD_WIN32_)
 struct TinyCThreadTSSData {
@@ -689,15 +639,6 @@ thrd_t thrd_current(void)
 #endif
 }
 
-int thrd_is_current(thrd_t thr) {
-#if defined(_TTHREAD_WIN32_)
-	return GetThreadId(thr) == GetCurrentThreadId();
-#else	
-	return (pthread_self() == thr);
-#endif
-}
-
-
 int thrd_detach(thrd_t thr)
 {
   thrd_is_detached = 1;
@@ -808,14 +749,6 @@ void thrd_yield(void)
 #else
   sched_yield();
 #endif
-}
-
-int thrd_setname (const char *name) {
-#if HAVE_PTHREAD_SETNAME_GNU
-  if (!pthread_setname_np(pthread_self(), name))
-    return thrd_success;
-#endif
-  return thrd_error;
 }
 
 int tss_create(tss_t *key, tss_dtor_t dtor)
@@ -998,51 +931,9 @@ void call_once(once_flag *flag, void (*func)(void))
 #endif /* defined(_TTHREAD_WIN32_) */
 
 
-#if !defined(_TTHREAD_WIN32_)
-int rwlock_init (rwlock_t *rwl) {
-        int r = pthread_rwlock_init(rwl, NULL);
-        if (r) {
-                errno = r;
-                return thrd_error;
-        }
-        return thrd_success;
-}
-
-int rwlock_destroy (rwlock_t *rwl) {
-        int r = pthread_rwlock_destroy(rwl);
-        if (r) {
-                errno = r;
-                return thrd_error;
-        }
-        return thrd_success;
-}
-
-int rwlock_rdlock (rwlock_t *rwl) {
-        int r = pthread_rwlock_rdlock(rwl);
-        assert(r == 0);
-        return thrd_success;
-}
-
-int rwlock_wrlock (rwlock_t *rwl) {
-        int r = pthread_rwlock_wrlock(rwl);
-        assert(r == 0);
-        return thrd_success;
-}
-
-int rwlock_rdunlock (rwlock_t *rwl) {
-        int r = pthread_rwlock_unlock(rwl);
-        assert(r == 0);
-        return thrd_success;
-}
-
-int rwlock_wrunlock (rwlock_t *rwl) {
-        int r = pthread_rwlock_unlock(rwl);
-        assert(r == 0);
-        return thrd_success;
-}
-
-#endif /* !defined(_TTHREAD_WIN32_) */
 
 #ifdef __cplusplus
 }
 #endif
+
+#endif /* !WITH_C11THREADS */
