+++ tests/CMakeLists.txt
@@ -80,13 +80,13 @@ set(
     0088-produce_metadata_timeout.c
     8000-idle.cpp
     test.c
-    testcpp.cpp    
+    testcpp.cpp
 )
 
 if(NOT WIN32)
     list(APPEND sources sockem.c)
 else()
-    list(APPEND sources ../src/tinycthread.c)
+    list(APPEND sources ../src/tinycthread.c ../src/tinycthread_extra.c)
 endif()
 
 add_executable(rdkafka_test ${sources})
